RavenDB-13107 - refactoring
