+++ src/Raven.Server/Documents/Handlers/RachisDatabaseHandler.cs
@@ -7,8 +7,8 @@ namespace Raven.Server.Documents.Handlers
 {
     public class RachisDatabaseHandler : DatabaseRequestHandler
     {
-        [RavenAction("/databases/*/wait-for-indexes-notification", "POST", AuthorizationStatus.Operator)]
-        public async Task WaitForIndexesNotification()
+        [RavenAction("/databases/*/wait-for-raft-commands", "POST", AuthorizationStatus.Operator)]
+        public async Task WaitForRaftCommands()
         {
             using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
             {
